# api.yaml
openapi: 3.0.3
info:
  title: Product Scanner API
  description: |
    상품 정보 스캔 및 업데이트 워크플로우 API

    ## 지원 플랫폼
    | 플랫폼 | 스캔 방식 | URL 패턴 |
    |--------|----------|----------|
    | hwahae | HTTP API | hwahae.co.kr |
    | musinsa | HTTP API | musinsa.com |
    | zigzag | GraphQL | zigzag.kr |
    | oliveyoung | Playwright | oliveyoung.co.kr |
    | kurly | Playwright | kurly.com |
    | ably | Playwright | a-bly.com |

    ## 워크플로우 종류
    - **Platform Update**: 플랫폼별 대량 상품 업데이트 (스케줄러 사용)
    - **Platform Validation**: 플랫폼별 대량 상품 검증 (업데이트 없음)
    - **Extract ProductSet**: 개별 ProductSet ID 기반 추출
    - **Extract URL**: URL 기반 단일 상품 추출

  version: 2.0.0
  contact:
    name: Scoob Scraper Team
servers:
  - url: http://localhost:3989
    description: 로컬 개발 서버
  - url: http://product_scanner:3000
    description: Docker 내부 통신
...
paths:
...
  /api/v2/llm/generate-description:
    post:
      tags: [LLM]
      summary: 상품 설명 및 카테고리 생성
      description: |
        URL Context를 활용하여 상품 페이지를 분석하고
        **마케팅용 상품 설명**과 **카테고리**를 생성합니다.

        **동기 API** - 결과를 즉시 반환합니다 (10~30초 소요).

        ## 2단계 LLM 호출 구조
        Gemini 2.5 Flash에서 URL Context와 Structured Output을 동시 사용할 수 없어
        2단계로 분리하여 호출합니다.

        | 단계 | 목적 | 입력 | 출력 |
        |------|------|------|------|
        | 1단계 | URL Context로 정보 추출 | URL 목록 | 마케팅 문구, 효능, 카테고리 정보 (텍스트) |
        | 2단계 | Structured Output 생성 | 추출된 정보 | 상품 설명, 카테고리 (JSON) |

        ## 상품 설명 스타일
        - 짧고 임팩트 있는 마케팅 카피 스타일
        - "~입니다" 같은 문장 종결 금지
        - 핵심 효능/특징을 압축하여 전달

        **예시**:
        - "내 피부인듯 맑고 투명하게 미백 톤업크림"
        - "도톰한 당근 한 장으로 덮는 열감 진정"
        - "저분자 히알루론산이 피부 깊숙이 수분 채우는 고보습 세럼"

        ## LLM 비용
        - 모델: `gemini-2.5-flash`
        - 평균 비용: ~$0.002/요청 (약 ₩3)
        - 비용 로그: `results/{date}/llm_cost__{date}.jsonl`
          - platform: `description` (URL Context 추출)
          - platform: `description` (Structured Output)

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateDescriptionRequest'
            examples:
              basic:
                summary: 기본 사용
                value:
                  brand: "토리든"
                  product_name: "다이브인 저분자 히알루론산 세럼"
                  urls:
                    - "https://www.oliveyoung.co.kr/store/goods/getGoodsDetail.do?goodsNo=A000000238213"
                    - "https://zigzag.kr/catalog/products/131281148"
              single_url:
                summary: 단일 URL
                value:
                  brand: "라네즈"
                  product_name: "워터뱅크 블루 히알루로닉 크림"
                  urls:
                    - "https://www.oliveyoung.co.kr/store/goods/getGoodsDetail.do?goodsNo=A000000231822"
      responses:
        '200':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateDescriptionResponse'
              examples:
                success:
                  summary: 성공 응답
                  value:
                    success: true
                    data:
                      description: "저분자 히알루론산이 피부 깊숙이 수분 채우는 고보습 세럼"
                      category:
                        id: 5
                        path: "스킨케어 > 에센스/세럼/앰플"
                      usage:
                        stage1:
                          input: 180
                          output: 463
                          url_context: 12587
                        stage2:
                          input: 2148
                          output: 77
                        total:
                          input: 2328
                          output: 540
                          url_context: 12587
                          total: 15455
                          cost_usd: 0.002561
                      model: "gemini-2.5-flash"
                      duration_ms: 6788
        '400':
          description: 잘못된 요청
          content:
            application/json:
              example:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "brand is required"
        '500':
          description: LLM API 오류
          content:
            application/json:
              examples:
                rate_limit:
                  summary: Rate Limit 초과
                  value:
                    success: false
                    error:
                      code: "RATE_LIMIT_EXCEEDED"
                      message: "Gemini API rate limit exceeded"
                internal:
                  summary: 내부 오류
                  value:
                    success: false
                    error:
                      code: "INTERNAL_ERROR"
                      message: "서버 내부 오류가 발생했습니다"
  /api/v2/llm/cost-stats:
    get:
      tags: [LLM]
      summary: 오늘 LLM 비용 통계
      description: |
        오늘 사용된 LLM 비용 통계를 조회합니다.

        **포함 정보**:
        - 총 비용 (USD)
        - 총 레코드 수
        - 입력/출력 토큰 수
        - 작업별/플랫폼별 비용 분류
      responses:
        '200':
          description: 통계 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmCostStatsResponse'
              example:
                success: true
                data:
                  total_cost_usd: 0.123
                  total_records: 45
                  total_input_tokens: 12345
                  total_output_tokens: 6789
                  by_operation:
                    product_description_extract: 0.05
                    product_description_structured: 0.03
                    product_filtering: 0.04
                  by_platform:
                    "description": 0.05
                    "filtering": 0.05
        '500':
          description: 서버 오류
          content:
            application/json:
              example:
                success: false
                error:
                  code: "INTERNAL_ERROR"
                  message: "비용 통계 조회에 실패했습니다"
...
components:
  schemas:
...
    GenerateDescriptionRequest:
      type: object
      required: [brand, product_name, urls]
      properties:
        brand:
          type: string
          minLength: 1
          description: 브랜드명
        product_name:
          type: string
          minLength: 1
          description: 상품명
        urls:
          type: array
          items:
            type: string
            format: uri
          minItems: 1
          maxItems: 20
          description: |
            상품 페이지 URL 목록 (1~20개)

            URL Context로 분석할 상품 페이지들
    GenerateDescriptionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            description:
              type: string
              description: 생성된 마케팅용 상품 설명
            category:
              type: object
              properties:
                id:
                  type: integer
                  description: 카테고리 ID
                path:
                  type: string
                  description: 카테고리 경로 (예- "스킨케어 > 에센스/세럼/앰플")
            usage:
              type: object
              properties:
                stage1:
                  type: object
                  description: 1단계 (URL Context 추출) 토큰 사용량
                  properties:
                    input:
                      type: integer
                    output:
                      type: integer
                    url_context:
                      type: integer
                      description: URL Context에서 추출된 토큰 수
                stage2:
                  type: object
                  description: 2단계 (Structured Output) 토큰 사용량
                  properties:
                    input:
                      type: integer
                    output:
                      type: integer
                total:
                  type: object
                  description: 총 토큰 사용량
                  properties:
                    input:
                      type: integer
                    output:
                      type: integer
                    url_context:
                      type: integer
                    total:
                      type: integer
                    cost_usd:
                      type: number
                      format: float
                      description: 총 비용 (USD)
            model:
              type: string
              description: 사용된 모델명
            duration_ms:
              type: integer
              description: 처리 시간 (ms)
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
              description: 에러 코드
            message:
              type: string
              description: 에러 메시지
    LlmCostStatsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            total_cost_usd:
              type: number
              format: float
              description: 오늘 총 LLM 비용 (USD)
            total_records:
              type: integer
              description: 오늘 총 LLM 호출 수
            total_input_tokens:
              type: integer
              description: 오늘 총 입력 토큰 수
            total_output_tokens:
              type: integer
              description: 오늘 총 출력 토큰 수
            by_operation:
              type: object
              additionalProperties:
                type: number
              description: 작업별 비용 분류
            by_platform:
              type: object
              additionalProperties:
                type: number
              description: 플랫폼별 비용 분류

# ============================================================================
# Shell 스크립트 사용법
# ============================================================================
#
# ## 스케줄러 제어 (scheduler-control.sh)
#
# ```bash
# # 상태 확인
# ./scripts/scheduler-control.sh status
#
# # 스케줄러 시작
# ./scripts/scheduler-control.sh start
#
# # 스케줄러 중지
# ./scripts/scheduler-control.sh stop
#
# # 스케줄러 중지 + 대기 Job 삭제
# ./scripts/scheduler-control.sh stop --clear-queue
#
# # 원격 서버
# API_URL=http://remote-server:3989 ./scripts/scheduler-control.sh status
# ```
#
# ## 실행 중인 Job 확인 (check-running-jobs.sh)
#
# ```bash
# ./scripts/check-running-jobs.sh
#
# # 원격 서버
# API_URL=http://remote-server:3989 ./scripts/check-running-jobs.sh
# ```
#
# ## Worker 제어 (worker-control.sh)
#
# ```bash
# # Worker 상태 확인
# ./scripts/worker-control.sh status
#
# # 특정 Worker 재시작
# ./scripts/worker-control.sh restart oliveyoung
# ./scripts/worker-control.sh restart hwahae
#
# # 모든 Worker 재시작
# ./scripts/worker-control.sh restart all
#
# # 원격 서버
# API_URL=http://remote-server:3989 ./scripts/worker-control.sh status
# ```
#
# ## Alert Watcher 제어 (alert-watcher-control.sh)
#
# 테이블 모니터링 서비스 제어 스크립트 (collabo_banners, votes, pick_sections)
#
# ```bash
# # 상태 확인
# ./scripts/alert-watcher-control.sh status
#
# # Alert Watcher 시작
# ./scripts/alert-watcher-control.sh start
#
# # Alert Watcher 중지
# ./scripts/alert-watcher-control.sh stop
#
# # Alert Watcher 중지 + 대기 Job 삭제
# ./scripts/alert-watcher-control.sh stop --clear-queue
#
# # 원격 서버
# API_URL=http://remote-server:3989 ./scripts/alert-watcher-control.sh status
# ```
#
# ### 개별 모니터 테스트
# ```bash
# # Collabo Banner 모니터 테스트
# ./scripts/test-collabo-banner-monitor.sh
#
# # Votes 모니터 테스트
# ./scripts/test-votes-monitor.sh
#
# # Pick Sections 모니터 테스트
# ./scripts/test-pick-sections-monitor.sh
# ```
#
# ## 검색 API 스크립트
#
# ### 통합 검색 (6개 플랫폼 병렬)
# ```bash
# # 기본 사용법
# ./scripts/search-unified.sh "브랜드명" "상품명" [limit]
#
# # 예시
# ./scripts/search-unified.sh "라네즈" "워터뱅크 크림" 5
# ./scripts/search-unified.sh "토리든" "" 3
#
# # 원격 서버
# API_URL=http://remote-server:3989 ./scripts/search-unified.sh "토리든" "세럼" 5
# ```
#
# ### 단일 플랫폼 검색 (디버깅용)
# ```bash
# # 지그재그 검색 (curl 직접 호출)
# ./scripts/search-zigzag.sh "수분크림" 10
# ./scripts/search-zigzag.sh "토리든" 10 --json
#
# # 올리브영 검색 (Playwright)
# ./scripts/search-oliveyoung.sh "토리든" 10
#
# # 무신사 검색 (Playwright + Stealth)
# ./scripts/search-musinsa.sh "토리든" 10
#
# # 에이블리 검색 (Playwright + Stealth)
# ./scripts/search-ably.sh "토리든" 10
#
# # 컬리 검색 (Playwright + Stealth)
# ./scripts/search-kurly.sh "토리든" 10
#
# # 화해 검색 (Playwright + DOM 파싱)
# ./scripts/search-hwahae.sh "토리든" 10
# ```
#
# ## 워크플로우 테스트 스크립트
#
# ### Platform 워크플로우
# ```bash
# # 화해 업데이트 테스트
# LIMIT=5 SALE_STATUS=on_sale ./scripts/test-hwahae-update.sh
#
# # 무신사 검증 테스트 (업데이트 없음)
# LIMIT=10 ./scripts/test-musinsa-validation.sh
#
# # 올리브영 업데이트
# LIMIT=5 ./scripts/test-oliveyoung-update.sh
#
# # 지그재그 업데이트
# LIMIT=5 ./scripts/test-zigzag-update.sh
#
# # 컬리 업데이트
# LIMIT=5 ./scripts/test-kurly-update.sh
#
# # 에이블리 업데이트
# LIMIT=5 ./scripts/test-ably-update.sh
# ```
#
# ### Extract 워크플로우
# ```bash
# # ProductSet ID 기반 추출 + 업데이트
# ./scripts/test-extract-product-set-update.sh "product-set-uuid"
#
# # ProductSet ID 기반 검증
# ./scripts/test-extract-product-set-validation.sh "product-set-uuid"
#
# # URL 기반 추출
# ./scripts/test-extract-url-validation.sh "https://www.oliveyoung.co.kr/..."
# ```
#
# ### 환경변수
# | 변수 | 설명 | 기본값 |
# |------|------|--------|
# | API_URL | API 서버 주소 | http://localhost:3989 |
# | LIMIT | 조회할 상품 수. **생략 시 전체 조회** | 5 (테스트 스크립트) |
# | BATCH_SIZE | 배치 크기 | 10 |
# | CONCURRENCY | 동시 처리 수 | 1 (API), 4 (Playwright) |
# | WAIT_TIME_MS | Rate limiting 대기 시간 | 2500 |
# | SALE_STATUS | 대상 판매 상태 | on_sale |
# | UPDATE_SALE_STATUS | 판매 상태 업데이트 여부 | true |
#
# **주의**: 스케줄러는 limit 없이 실행되어 **전체 상품을 자동 pagination으로 조회**합니다.
# 테스트 시에만 LIMIT 환경변수를 사용하세요.
#
# ============================================================================